pyproject.toml
[tool.poetry]
name = "letras"
version = "1.0.0"
description = "Gospel lyrics scraper from letras.mus.br"
authors = ["Daniel Amaral <silva.daniel86@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
rich = "^13.7.0"
aiohttp = "^3.9.1"
asyncpg = "^0.29.0"
beautifulsoup4 = "^4.12.2"
lingua-language-detector = "^2.0.2"
tenacity = "^8.2.3"
pydantic = "^2.10.4"
pydantic-settings = "^2.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-cov = "^4.1.0"
responses = "^0.24.1"
black = "^23.12.0"
isort = "^5.13.2"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3